R script for modeling of carbon isotope alteration
This R script contains code for the construction of reactive-transport models (after work of Soetaert), these reactive-transport models construct diagenetic depth trajectories of solutes (porewater) and solids (sedimentation). A sensitivity experiment is constructed to determine the effect of several environmental parameters on the carbon isotope composition of bulk rock due to ongoing dissolution and recrystallization with depth, and near instantaneous authigenic cementation.  A virtual timeseries is simulated to explain residual carbon isotope variability and temporal trends in the amplitude of these variations as seen in Permian-Triassic rock sequences.